+++ src/lib/pages/contract-details/data.ts
@@ -38,7 +38,9 @@ export const useContractData = (
   const { data: contractDetail, isLoading: isContractDetailLoading } =
     useContractDetailByContractAddress(contractAddress);
 
-  const { assetInfos, isLoading: isAssetInfosLoading } = useAssetInfos();
+  const { assetInfos, isLoading: isAssetInfosLoading } = useAssetInfos({
+    withPrices: true,
+  });
   const { data: contractBalances, isLoading: isContractBalancesLoading } =
     useQuery(
       ["query", "contractBalances", balancesApiRoute, contractAddress],